Creation of a File Example.
Check that we have a DataStore.cs file, which creates the interface to Get and Set Data from a consumer. Later, we do an Adapter, that is called Implementation, which we create a class called Data, which is the one we are going to save.
Later, I need a Consumer,  To test things.
This creates a File called "Data1"