Destroying the Proyectile
Implementing a coroutine  for destroying the projectiles. Also, I'm Introducing a new Pattern: Mediator Pattern, that will modify this Ship class, making it a little more simple by creating new classes for the different behaviours the Ship will have, such as, Movement, shooting, Life Mechanics, and the mechanics that we will be joining altogether to make the ship alive.